moved sso user delete method to account controller
